feat: board 폴더 캡슐화 진행
