+++ app/components/Board/index.ts
@@ -1,0 +1,4 @@
+import BoardItem from "./BoardItem";
+import BoardWrapper from "./BoardWrapper";
+
+export { BoardItem, BoardWrapper };
